fix: use float64 for K_C eigendecomposition to handle repeated eigenvalues
When C > marker_embed_dim (32), E_norm @ E_norm.T is rank-deficient with
C-D eigenvalues exactly at marker_jitter. LAPACK's divide-and-conquer eigh
fails with 'too many repeated eigenvalues' in float32. Casting K_C to float64
before eigh and back to float32 after resolves convergence reliably with
negligible overhead (C is at most ~40, a tiny matrix).

+++ multiplex_model/modules/gp_covariance.py
@@ -523,7 +523,12 @@ class KroneckerMarkerCovariance(nn.Module):
         E = nn.functional.normalize(E, p=2, dim=1)
         C = E.shape[0]
         K_C = E @ E.T + self.marker_jitter * torch.eye(C, device=E.device, dtype=E.dtype)
-        lam_C, V_C = torch.linalg.eigh(K_C)
+        # Use float64 for eigh: when C > marker_embed_dim, K_C has C-D repeated eigenvalues
+        # at exactly marker_jitter. LAPACK's divide-and-conquer fails on near-repeated
+        # eigenvalues in float32; float64 precision resolves convergence reliably.
+        lam_C, V_C = torch.linalg.eigh(K_C.double())
+        lam_C = lam_C.to(E.dtype)
+        V_C = V_C.to(E.dtype)
 
         # triple_eigs[i, j, k] = kron_eigs[i,j] * lam_C[k] + kernel_jitter
         triple_eigs = self.kron_eigs.unsqueeze(-1) * lam_C.unsqueeze(0).unsqueeze(0) + self.kernel_jitter
